+++ src/apswversion.h
@@ -1,1 +1,1 @@
-#define APSW_VERSION "3.7.0.1-r1"
+#define APSW_VERSION "3.7.1-r1"